Add files via upload
The Linear and Logistic Regression algorithms are statistical techniques used to analyze and predict the likelihood of an outcome. In this case, they are being used to predict the total sales and the likelihood of a video game achieving sales in excess of $1 million. The algorithms take in a set of independent variables, such as the platform, genre, and developer, and the dependent variable, which is the actual sales of the game. It then builds a model based on these variables to predict the likelihood of a game achieving sales over a certain threshold.

The steps in the algorithm are:

1. Collect the necessary data: the sales data for each video game must be collected, which includes the platform, genre, developer, and game title.

2. Train the model: The model is trained using this data. The model uses the independent variables to predict the outcome, which in this case is the total sales and whether or not a video game will achieve sales in excess of $1 million.

3. Predict: Once the model is trained, it can be used to predict the sales of the game and the likelihood of a video game achieving sales in excess of $1 million.